src/requirements.txt
django==2.2.13
djangorestframework==3.12.4

django-allauth==0.41.0
django-celery-beat==2.1.0
django-celery-model==0.2.1
django-celery-results==1.1.2
django_celery_monitor==1.1.2
django-constance[database]==2.7.0
django-crispy-forms==1.8.1
django-extensions==2.2.6
django-filter==2.2.0
django-forms-builder==0.14.0
django-jazzmin==2.4.8
django-json-widget==1.0.0
django-modeladmin-reorder==0.3.1
django-polymorphic==2.1.2
django-redis==4.11.0
django-report-builder==6.3.2
django-rest-swagger==2.2.0
django-rest-polymorphic==0.1.1
django-simple-history==2.8.0
django-sql-explorer==2.1.2
django-user-tasks==2.1.0
django-task-api==1.1.6
django-datatables-view==1.19.1
django-notifications-hq==1.6.0
django-admin-tools==0.9.0
django-timezone-field==4.0
django-debug-toolbar==3.2
django-fsm==2.7.1
django-fsm-log==2.0.1
django-fsm-admin==1.2.4
django-admin-honeypot==1.1.0
drf-action-serializer==1.0.5
drf-extensions==0.7.1
drf-access-policy==1.1.0
django-csp==3.7

aiohttp==2.0.1
boto==2.49.0
bokeh==1.4.0
celery[redis]==4.4.7
cryptography==3.3.2
channels==2.1.2
channels-redis==2.4.2
daphne==2.2.1
dictor==0.1.2
graphviz==0.13.0
hvac==0.10.5
ipython==7.11.1
jedi==0.17.2 # https://github.com/ipython/ipython/issues/12753
jsonpath-ng==1.4.3
jupyter==1.0.0
lxml==4.6.2
markdown==3.1.1
matplotlib==3.1.2
networkx==2.5.1
numpy==1.18.1
nxpd==0.2.0
pandas==0.25.3
parse==1.14.0
profilehooks==1.11.0
psycopg2-binary==2.8.4
pydotplus==2.0.2
pycryptodome==3.10.1
pyquery==1.2.10
pygraphviz==1.5
redis==3.3.11
rules==2.2
textblob==0.15.3
tornado==5.1.1 # Jupyter kernel fixdjango-forms-builder
watchdog==0.9.0
whitenoise==5.0.1
pylint==2.5.3
python-ldap==3.3.0
uWSGI==2.0.17
rejson==0.5.4
pysnow==0.7.17
tabulate==0.8.9

pytest-django==4.1.0
pytest-cov
pytest-celery
pytest-benchmark
factory-boy==3.2.1
